backend/src/requirements.txt
concurrent-log-handler==0.9.16
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-JWT-Extended==3.24.0
Flask-Migrate==2.5.2
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
marshmallow==3.2.1
PyMySQL==0.9.3
pytest==5.2.1
requests==2.22.0